Move title to the components catalog
